pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agent-nebula"
version = "0.1.0"
description = "Universal infinite-loop agent workflow powered by Claude Code SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-code-sdk>=0.0.25",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
agent-nebula = "agent_nebula.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
